renamed navbar to navbar2 to prepare for rewriting
